# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  HTTP400BadRequest:
    properties:
      err_code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  HTTP404NotFound:
    properties:
      err_code:
        example: 404
        type: integer
      message:
        example: Not Found
        type: string
    type: object
  HTTP500InternalServerError:
    properties:
      err_code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.ClusterBootstrapConfig:
    properties:
      ccvm:
        description: ccvm bootstrap flag
        example: "true"
        type: string
      gfs_configure:
        description: gfs configure flag
        example: "true"
        type: string
      local_configure:
        description: local configure flag
        example: "false"
        type: string
      scvm:
        description: scvm bootstrap flag
        example: "false"
        type: string
      wall:
        description: wall configuration flag
        example: "true"
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.ClusterCCVMConfig:
    properties:
      ip:
        description: ccvm IP
        example: 10.10.31.10
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.ClusterConfigSection:
    properties:
      backup_path:
        description: backup path for ccvm
        example: /mnt/glue-gfs/backup/ccvm
        type: string
      ccvm:
        allOf:
        - $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterCCVMConfig'
        description: ccvm config
      external_timeserver:
        description: external timeserver
        example: time.google.com
        type: string
      hosts:
        description: host list
        items:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterHost'
        type: array
      iscsi_storage:
        description: iscsi storage usage
        example: "false"
        type: string
      mngtNic:
        allOf:
        - $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterMngtNicConfig'
        description: management NIC config
      pcsCluster:
        allOf:
        - $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterPCSClusterConfig'
        description: pcs cluster IPs
      type:
        description: cluster type
        example: ablestack-vm
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.ClusterHost:
    properties:
      ablecube:
        example: 10.10.31.1
        type: string
      ablecubePn:
        example: 100.100.31.1
        type: string
      hostname:
        example: ablecube31-1
        type: string
      index:
        example: "1"
        type: string
      scvm:
        example: 100.100.31.11
        type: string
      scvmCn:
        example: 100.200.31.11
        type: string
      scvmMngt:
        example: 10.10.31.11
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.ClusterLicenseConfig:
    properties:
      status:
        description: license status
        example: "true"
        type: string
      type:
        description: license type
        example: ablestack
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.ClusterMngtNicConfig:
    properties:
      cidr:
        description: management NIC CIDR
        example: "16"
        type: string
      dns:
        description: management NIC DNS
        example: 8.8.8.8
        type: string
      gw:
        description: management NIC gateway
        example: 10.10.0.1
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.ClusterPCSClusterConfig:
    properties:
      hostname1:
        description: 'pcs cluster node #1'
        example: 10.10.31.1
        type: string
      hostname2:
        description: 'pcs cluster node #2'
        example: 10.10.31.2
        type: string
      hostname3:
        description: 'pcs cluster node #3'
        example: 10.10.31.3
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.ClusterSecurityPatchConfig:
    properties:
      status:
        description: security patch status
        example: "false"
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.ClusterSystemProfile:
    properties:
      bootstrap:
        allOf:
        - $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterBootstrapConfig'
        description: bootstrap flags
      license:
        allOf:
        - $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterLicenseConfig'
        description: license info
      security_patch:
        allOf:
        - $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterSecurityPatchConfig'
        description: security patch status
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.GlueClusterUpdateRequest:
    properties:
      action:
        description: 'action: set_noout/unset_noout'
        example: set_noout
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.GlueClusterUpdateResponse:
    properties:
      action:
        example: set_noout
        type: string
      code:
        example: 200
        type: integer
      message:
        example: success
        type: string
      retname:
        example: Maintenance Mode On
        type: string
      val:
        example: success
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.GlueConfigUpdateResponse:
    properties:
      action:
        example: update
        type: string
      code:
        example: 200
        type: integer
      message:
        example: Glue Config All cube host and scvm update Success
        type: string
      results:
        items:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.GlueConfigUpdateTargetResult'
        type: array
      source:
        example: local
        type: string
      val: {}
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.GlueConfigUpdateTargetResult:
    properties:
      code:
        example: 200
        type: integer
      hostname:
        example: ablecube31-1
        type: string
      message:
        example: ok
        type: string
      role:
        example: ablecube
        type: string
      step:
        example: copy
        type: string
      target:
        example: 10.10.31.1
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.SystemConfigRequest:
    properties:
      action:
        description: 'action: status/update/allUpdate/reset'
        example: status
        type: string
      depth1:
        description: depth1 key (bootstrap/license/security_patch)
        example: bootstrap
        type: string
      depth2:
        description: depth2 key (e.g. scvm/ccvm/wall/gfs_configure/local_configure/status/type)
        example: scvm
        type: string
      option:
        description: 'option: all (fan-out to all ablecube hosts)'
        example: all
        type: string
      value:
        description: value for update
        example: "true"
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.SystemConfigResponse:
    properties:
      code:
        example: 200
        type: integer
      val: {}
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.VersionUpdateRequest:
    properties:
      action:
        description: 'action: info/run'
        example: info
        type: string
      mount_path:
        description: mounted ABLESTACK ISO path
        example: /mnt/ablestack-iso
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.VersionUpdateResponse:
    properties:
      action:
        example: info
        type: string
      code:
        example: 200
        type: integer
      message:
        type: string
      retname:
        type: string
      val: {}
    type: object
...
info:
  contact:
    email: ycyun@ablecloud.io
    name: API Support
    url: https://www.ablecloud.io/support
  description: This is a Cube-API server.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://ablecloud.io/
  title: Cube API
  version: "1.0"
paths:
...
  /cube/cluster/config:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: cluster.json의 clusterConfig만 반환합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterConfigSection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Show Cluster Config
      tags:
      - CUBE - Cluster
...
  /cube/glue/config/update:
    post:
      consumes:
      - application/json
      description: /etc/ceph 설정을 생성하거나 pcsCluster.hostnameN에서 가져온 뒤 cluster.json hosts[].ablecube/scvmMngt
        대상으로 배포합니다. 요청 body 없이 호출합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.GlueConfigUpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Update Glue Config
      tags:
      - CUBE - Glue Config
...
  /cube/gluecluster/update:
    post:
      consumes:
      - application/json
      description: 스토리지 클러스터 유지보수 모드를 설정하거나 해제합니다.
      parameters:
      - description: glue cluster update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.GlueClusterUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.GlueClusterUpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Glue Cluster Update
      tags:
      - CUBE - Glue Cluster
...
  /cube/system/config:
    get:
      consumes:
      - application/json
      description: systemProfile을 반환합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterSystemProfile'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: System Config
      tags:
      - CUBE - System
    post:
      consumes:
      - application/json
      description: systemProfile을 상태 조회/수정합니다.
      parameters:
      - description: system config request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.SystemConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.SystemConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: System Config Update
      tags:
      - CUBE - System
...
  /cube/version/update:
    post:
      consumes:
      - application/json
      description: 마운트된 ABLESTACK ISO의 버전 정보를 조회하거나 update.sh를 실행합니다.
      parameters:
      - description: version update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.VersionUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.VersionUpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: ABLESTACK Version Update
      tags:
      - CUBE - Version
...
  /glue:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: GLUE의 상태값을 보여줍니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TypeGlueStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Show Status of GLUE
      tags:
      - API
      - Glue
      - GLUE
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Enter the token with the Bearer prefix, e.g. Bearer eyJ...
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"